requirements.txt
argparse==1.2.1
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
blinker==1.3
decorator==3.4.0
gunicorn==19.1.1
itsdangerous==0.24
pbr==0.10.3
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
PyMySQL==0.6.3
wsgiref==0.1.2
